% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.R
\docType{package}
\name{tandemrng-package}
\alias{tandemrng}
\alias{tandemrng-package}
\title{tandemrng: Tandem8x32 Pseudorandom Number Generator}
\description{
A noncryptographic pseudorandom number generator, Tandem8x32, built to be fast on CPUs and GPUs alike. The package wraps the reference C implementation and produces the same stream, bit for bit, as the other implementations of the specification \url{https://github.com/tandem-rng/spec}. It provides generators with cheap splitting by index, by purpose and by fork, fills of uniform doubles and of unsigned words, and a hook that makes Tandem the base R generator through RNGkind("user-supplied").
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tandem-rng/tandem-r}
  \item \url{https://github.com/tandem-rng/spec}
  \item Report bugs at \url{https://github.com/tandem-rng/tandem-r/issues}
}

}
\author{
\strong{Maintainer}: Jessica Cox \email{jmcox@posteo.de}

Authors:
\itemize{
  \item Jessica Cox \email{jmcox@posteo.de}
}

}
\keyword{internal}
