% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{download_TanDEM}
\alias{download_TanDEM}
\title{Download TanDEM-X}
\usage{
download_TanDEM(
  lon = c(5, 16),
  lat = c(45, 55),
  usr = getOption("geoservice.usr"),
  srv = "geoservice.dlr",
  dstdir = "TanDEM-X"
)
}
\arguments{
\item{lon}{numeric vector of length two indicating the longitude range.}

\item{lat}{numeric vector of length two indicating the latitude range.}

\item{usr}{character. username.}

\item{srv}{character. service.}

\item{dstdir}{character. destination directory.}
}
\description{
A download tool to retrieve TanDEM-X data.
}
\details{
Download of tiles using \code{wget} from https.
A character vector of the downloaded tif files is returned.
}
\examples{
\dontrun{
x <- download_TanDEM(lon = 11.5, lat = 47.5, dstdir = "test")
}
}
\keyword{data}
